Greatly reduce memory usage during training
Previously, when subtasks need to share some data, we let their parents own it, and only free it when all the subtasks are joined. However, since subtasks can spawn sub-subtasks and so on, such data can stay in memory long after they are useful.

Now we let subtasks share the ownership of such data, and a subtask can release its handle to the data as soon as it no longer needs it. This way, we can free the memory occupied by unneeded data ASAP, greatly reducing memory usage at the cost of incurring the overhead of smart pointers with (atomic) ref counting.