Store classifier weights on each node as a matrix
For now we still use the CsMat implementation from the `sprs` library. Later we'll replace it with something more efficient for our purpose.

Here we also removed the logic where if all training examples are positive, we skip training the corresponding classifier. If I remember correctly, the logic was added only to speed up training for a little bit, so this hopefully shouldn't affect prediction performance.